% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny app}
\usage{
launch_app()
}
\value{
No return value, and called for side effects.
}
\description{
Starts the Germany HAI Explorer included with this package.
}
