% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{sim_weekly}
\alias{sim_weekly}
\title{Simulated German HAI weekly series, 2011–2012}
\format{
A tibble with columns:
\describe{
\item{date}{Date for ISO week start (Monday)}
\item{year}{Integer year}
\item{iso_week}{Integer ISO week number}
\item{hai}{Character infection type}
\item{cases_week}{Integer simulated cases}
\item{deaths_week}{Integer simulated deaths}
\item{dalys_week}{Numeric simulated DALYs}
}
}
\usage{
sim_weekly
}
\description{
Simulated German HAI weekly series, 2011–2012
}
\keyword{datasets}
