% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{sim_daily}
\alias{sim_daily}
\title{Simulated German HAI daily series, 2011–2012}
\format{
A tibble with columns:
\describe{
\item{date}{Date}
\item{year}{Integer year}
\item{yday}{Day-of-year (1–366)}
\item{wday}{Day-of-week (1=Mon, 7=Sun)}
\item{hai}{Character infection type}
\item{cases_day}{Integer simulated cases}
\item{deaths_day}{Integer simulated deaths}
\item{dalys_day}{Numeric simulated DALYs}
}
}
\usage{
sim_daily
}
\description{
Simulated German HAI daily series, 2011–2012
}
\keyword{datasets}
